requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
cloudpickle==2.2.1
cmake==3.26.3
decorator==5.1.1
dm-tree==0.1.8
filelock==3.12.0
flatbuffers==1.12
gast==0.3.3
google-auth==2.19.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.4
imageio==2.30.0
importlib-metadata==6.6.0
Jinja2==3.1.2
joblib==1.2.0
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==16.0.0
lit==16.0.5
Markdown==3.4.3
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.21.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
Pillow==9.5.0
pkg_resources==0.0.0
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.10.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.15.0
skia-python==87.5
sympy==1.12
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.1
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-probability==0.12.2
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
triton==2.0.0
typing-extensions==3.7.4.3
urllib3==1.26.16
Werkzeug==2.3.4
wrapt==1.12.1
zipp==3.15.0
seaborn==0.12.2